+++ GameEngine.Common/CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
 
 
 ### Other Changes
@@ -32,7 +32,7 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
 
 
 ### Other Changes
@@ -48,4 +48,4 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))