+++ GameEngine.Physics/CHANGELOG.md
@@ -31,7 +31,7 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
 
 ## [0.0.1](https://github.com/braybrandon/Game-Engine/compare/v0.0.1...v0.0.1) (2025-08-24)
 
@@ -57,7 +57,7 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
 
 ## [0.0.1](https://github.com/braybrandon/Game-Engine/compare/v0.1.3...v0.0.1) (2025-08-24)
 
@@ -83,4 +83,4 @@
 
 ### Code Refactoring
 
-* **engine:** clean up GameEngine.Engine — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))
+* **engine:** clean up GameEngine.Gameplay — restructure, docs, decouple ([58f8d48](https://github.com/braybrandon/Game-Engine/commit/58f8d481363d005af60855b5893917e9f35a133b))